MB-11629: Flag for session_ctr in cookie structure
Flag 'session_counted' added to the cookie structure.
When session counter is incremented for a cookie, this
flag will be set to true, and will be set to false for
the cookie again whenever the session counter is
decremented. This will ensure that threads will not
race in decrementing the session counter resulting in
an underflow.

Change-Id: Ia119c0d99a68b58bf2c5ec64a0b1393cdc2a6d3f